Terminate all JSON output with a newline
I think it would be nice if the output always ended with a newline. It's weird
when it doesn't because the next prompt starts on the same line as the last
line of output and that just doesn't look great. Also it's nice to have the
output end with a newline because then you can redirect to a file and have that
file adhere to POSIX standards - e.g.:

  - http://stackoverflow.com/questions/729692/why-should-text-files-end-with-a-newline

Fixes: GH-23